+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,10 +1,14 @@
 ## 0.0.10
 * Migrate from `dart:html` to `package:web`.
+* Add an example of registering a service extension and calling it from the
+DevTools extension.
+* Document the DevTools extension examples.
+* Add documentation to [ExtensionManager] public APIs.
 * Fix some bugs with the `build_and_copy` command for Windows.
 * Add an example `launch.json` file in the `example/foo` directory.
 * Clean up the package readme to make instructions Windows-compatible.
 * Update the README with instructions for joining the Flutter Discord server.
-* Bump `package:devtools_shared` dependency to ^6.0.0
+* Bump `package:devtools_shared` dependency to ^6.0.1
 * Bump `package:devtools_app_shared` dependency to ^0.0.7
 * Bump `package:vm_service` dependency to ^13.0.0.